Move notarization into a separate file
